fixed UI issues and made button work with SwiftUI and UIKit
